requirements.txt
Django==1.5.1
django-hstore==1.2.1
djangorestframework==2.3.12
